// client/src/interface/InputData.d.ts
export default interface InputData {
    PRIMARY_OFFENCE: string;
    OCC_YEAR: number;
    OCC_MONTH: string;
    OCC_DOW: string;
    OCC_DAY: number;
    OCC_DOY: number;
    OCC_HOUR: number;
    REPORT_YEAR: number;
    REPORT_MONTH: string;
    REPORT_DOW: string;
    REPORT_DAY: number;
    REPORT_DOY: number;
    REPORT_HOUR: number;
    DIVISION: string;
    LOCATION_TYPE: string;
    PREMISES_TYPE: string;
    BIKE_MAKE: string;
    BIKE_MODEL: string;
    BIKE_TYPE: string;
    BIKE_SPEED: number;
    BIKE_COLOUR: string;
    BIKE_COST: number;
    LONGITUDE: number;
    LATITUDE: number;
}